Embed flex layouts into each other

+++ app/src/main/java/com/robotjatek/wplauncher/Components/Button/Button.java
@@ -85,8 +85,8 @@ public class Button implements UIElement {
     }
 
     @Override
-    public Size<Float> measure() {
-        return new Size<>(0f, 100f); // TODO: configure height
+    public Size<Integer> measure() {
+        return new Size<>(0, 100); // TODO: configure height
     }
 
     public void setText(String text) {
