+++ app/src/main/java/com/robotjatek/wplauncher/Components/Label/Label.java
@@ -69,7 +69,7 @@ public class Label implements UIElement {
     }
 
     @Override
-    public Size<Float> measure() {
+    public Size<Integer> measure() {
         var paint = new Paint(Paint.ANTI_ALIAS_FLAG);
         paint.setTextAlign(Paint.Align.LEFT);
         paint.setTypeface(Typeface.create("sans-serif-light", _typeFace));
@@ -77,7 +77,7 @@ public class Label implements UIElement {
         var textWidth = paint.measureText(_text);
         var fm = paint.getFontMetrics();
         var textHeight = fm.descent - fm.ascent;
-        return new Size<>(textWidth, textHeight);
+        return new Size<>((int)textWidth, (int)textHeight);
     }
 
     public String getText() {
