+++ app/src/main/java/com/robotjatek/wplauncher/Components/UIElement.java
@@ -4,8 +4,8 @@ import com.robotjatek.wplauncher.IDrawContext;
 import com.robotjatek.wplauncher.QuadRenderer;
 
 public interface UIElement {
-     void draw(float[] proj, float[] view, IDrawContext<UIElement> drawContext, QuadRenderer renderer);
-     Size<Float> measure();
+     void draw(float[] proj, float[] view, IDrawContext<UIElement> drawContext, QuadRenderer renderer); // TODO: pass delta?
+     Size<Integer> measure();
      void dispose();
      void onTap();
 }
